+++ R/utils.R
@@ -1,3 +1,29 @@
+#' Check for Longitude/Latitude Coordinates
+#'
+#' This function checks whether a pair of coordinate vectors represent
+#' geographic longitude and latitude values. The function returns `TRUE`
+#' if all longitude values are within -180 (- tolerance) and 180 (+ tolerance)
+#' and the latitude is within -90 (- tolerance) and 90 (+ tolerance).
+#'
+#' @param lon Numeric vector of longitudes in degrees.
+#' @param lat Numeric vector of latitudes in degrees.
+#' @param tolerance Numeric tolerance (in degrees) for checking the global
+#'   longitude/latitude bounds. Default is 0.1.
+#'
+#' @return A logical value. `TRUE` if values are within the ranges and `FALSE` otherwise.
+#'
+#' @examples
+#' is_lonlat(lon = c(-3, 10, 179), lat = c(40, -20, 5))
+#' is_lonlat(lon = c(100000, 150000), lat = c(4500000, 4600000))
+#'
+#' @export
+is_lonlat <- function(lon, lat, tolerance = 0.1) {
+  lon_ok <- all(lon >= -180 - tolerance & lon <= 180 + tolerance, na.rm = TRUE)
+  lat_ok <- all(lat >= -90  - tolerance & lat <=  90 + tolerance, na.rm = TRUE)
+
+  return(lon_ok && lat_ok)
+}
+
 #' Convert Geographic Coordinates to Cartesian Coordinates
 #'
 #' This function converts geographic coordinates, given as longitude and latitude in degrees, to Cartesian coordinates (x, y, z) assuming a spherical Earth model.